:white_check_mark: Test: Move property checks to createMemoryClient.spec.ts
