+++ packages/memory-client/src/createMemoryClient.js
@@ -6,8 +6,10 @@ import { createClient, createTransport, publicActions, testActions } from 'viem'
 // TODO strongly type this! Currently it's return type is inferred
 
 /**
-* A local EVM instance running in JavaScript. Similar to Anvil in your browser
+* A local EVM instance running in JavaScript. Similar to Anvil 
 * It wraps the viem [public client](https://viem.sh/docs/clients/public#public-client) and [test client](https://viem.sh/docs/clients/test)
+* It also supports powerful [tevm actions](https://tevm.sh/learn/actions/) prefixed with `tevm`
+* _tevm has the internal api for low level control of the EVM
 * @param {import('@tevm/base-client').BaseClientOptions} [options]
 * @returns {import('./MemoryClient.js').MemoryClient}
 * @example
@@ -36,44 +38,45 @@ import { createClient, createTransport, publicActions, testActions } from 'viem'
 *  )
 *  console.log(balance) // 1n
 *  ```
+*  @see {@link https://tevm.sh/learn/client/} for more information on the `tevm` clients
+*  @see {@link https://tevm.sh/learn/actions/} for more information on the `tevm` actions
 */
 export const createMemoryClient = (options) => {
-	const tevm = createBaseClient(options)
-		.extend(tevmSend())
-		.extend(eip1993EventEmitter())
-		.extend(requestEip1193())
-		.extend(tevmActions())
-
-	return createClient({
-		type: 'tevm',
-		.../** @type any*/ (options?.common !== undefined ? { chain: options.common } : { chain: tevmDefault }),
-		transport: () =>
-			createTransport({
-				request: /** @type any*/ (tevm.request),
-				type: 'tevm',
-				name: /**options?.name ??*/ 'Tevm transport',
-				key: /*options?.key ??*/ 'tevm',
-				timeout: /*options?.timeout ?? */ 20_000,
-				retryCount: /*options?.retryCount ??*/ 3,
-				retryDelay: /* options?.retryDelay ?? */ 150,
-			}),
-	})
-		.extend(testActions({ mode: 'anvil' }))
-		.extend(publicActions)
-		.extend(() => {
-			return {
-				_tevm: tevm,
-				tevmReady: tevm.ready,
-				tevmCall: tevm.call,
-				tevmContract: tevm.contract,
-				tevmScript: tevm.script,
-				tevmDeploy: tevm.deploy,
-				tevmMine: tevm.mine,
-				tevmLoadState: tevm.loadState,
-				tevmDumpState: tevm.dumpState,
-				tevmSetAccount: tevm.setAccount,
-				tevmGetAccount: tevm.getAccount,
-				...(tevm.forkTransport !== undefined ? { forkTransport: tevm.forkTransport } : {}),
-			}
-		})
+const tevm = createBaseClient(options)
+.extend(tevmSend())
+.extend(eip1993EventEmitter())
+.extend(requestEip1193())
+.extend(tevmActions())
+return createClient({
+type: 'tevm',
+.../** @type any*/ (options?.common !== undefined ? { chain: options.common } : { chain: tevmDefault }),
+transport: () =>
+createTransport({
+request: /** @type any*/ (tevm.request),
+type: 'tevm',
+name: /**options?.name ??*/ 'Tevm transport',
+key: /*options?.key ??*/ 'tevm',
+timeout: /*options?.timeout ?? */ 20_000,
+retryCount: /*options?.retryCount ??*/ 3,
+retryDelay: /* options?.retryDelay ?? */ 150,
+}),
+})
+.extend(testActions({ mode: 'anvil' }))
+.extend(publicActions)
+.extend(() => {
+return {
+_tevm: tevm,
+tevmReady: tevm.ready,
+tevmCall: tevm.call,
+tevmContract: tevm.contract,
+tevmScript: tevm.script,
+tevmDeploy: tevm.deploy,
+tevmMine: tevm.mine,
+tevmLoadState: tevm.loadState,
+tevmDumpState: tevm.dumpState,
+tevmSetAccount: tevm.setAccount,
+tevmGetAccount: tevm.getAccount,
+...(tevm.forkTransport !== undefined ? { forkTransport: tevm.forkTransport } : {}),
+}
+})
 }
